fix: resolve constructor dependency qualifiers and map JPA grade columns to Short

+++ user-service/src/main/java/uz/quizplatform/userservice/infrastructure/persistence/entity/PaymentRequestJpaEntity.java
@@ -30,7 +30,7 @@ public class PaymentRequestJpaEntity {
     private String major;
 
     @Column(name = "grade")
-    private Integer grade;
+    private Short grade;
 
     @Column(name = "study_method")
     private String studyMethod;
