+++ user-service/src/main/java/uz/quizplatform/userservice/infrastructure/persistence/entity/UserJpaEntity.java
@@ -42,7 +42,7 @@ public class UserJpaEntity {
     private String major;
 
     @Column(name = "grade")
-    private Integer grade;
+    private Short grade;
 
     @Column(name = "study_method")
     @Enumerated(EnumType.STRING)
